squishy: Huge Renovation
This was done out of a frustration of where the code-base was and how hard it was to
make the needed changes to add rev2 support for things like the bootloader and such.

This was due to some very bad decisions that were made early on in a "just hack it together"
mindset early on to get something working, but the problems from that proliferated around
and made it very *very* painful to actually make the needed changes now.

As such, we basically took a hack-saw to the code and tried to improve everything we could see
that was wrong and hampering us. As a result of this we also improved the internal documentation
a bit, and also fixed some sneaky bugs we found, as well as improving the design of some things
such as the cache system.